requirements.txt
flask
requests
pillow
gunicorn
Jinja2
joblib
kiwisolver
MarkupSafe




pipreqs
pyparsing
python-dateutil
pytz
requests


six

threadpoolctl
urllib3
Werkzeug
yarg
